requirements.txt.txt
streamlit
opencv-python-headless
pillow
matplotlib
numpy